Calculadora alterando tipo da variável para float

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -7,17 +7,17 @@ public class Calculadora {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		
-		int a, b; 
+		float a, b; 
 		
 		System.out.println("Digite o primeiro valor: ");
 		a = scan.nextInt();
 		System.out.println("Digite o segundo valor: ");	
 		b = scan.nextInt();
 		
-		int soma = soma(a,b);
-		int subtracao = subtracao(a,b);
-		int divisao = divisao(a,b);
-		int multiplicacao = multiplicacao(a,b);
+		float soma = soma(a,b);
+		float subtracao = subtracao(a,b);
+		float divisao = divisao(a,b);
+		float multiplicacao = multiplicacao(a,b);
 		
 		System.out.println("Soma: " + soma);
 		System.out.println("Subtração: " + subtracao);
@@ -26,16 +26,16 @@ public class Calculadora {
 		
 	}
 
-	public static int soma(int a, int b) {
+	public static float soma(float a, float b) {
 		return a + b;
 	}
-	public static int subtracao(int a, int b) {
+	public static float subtracao(float a, float b) {
 		return a - b;
 	}
-	public static int divisao(int a, int b) {
+	public static float divisao(float a, float b) {
 		return a / b;
 	}
-	public static int multiplicacao(int a, int b) {
+	public static float multiplicacao(float a, float b) {
 		return a * b;
 	}
